requirements.txt
Flask
pymongo
Faker
mysql-connector-python
